+++ components/Editor.vue
@@ -94,8 +94,17 @@
     </div>
     <div class="mt-3">
       <b-button @click="clear">Clear</b-button>
-      <b-button class="float-right" variant="primary" @click="save"
-        >Save</b-button
+      <b-button
+        class="float-right"
+        variant="primary"
+        :disabled="loading"
+        @click="save"
+        ><v-progress-circular
+          v-show="loading"
+          indeterminate
+          size="24"
+          width="2"
+        /><span v-show="!loading">Save</span></b-button
       >
     </div>
   </div>
@@ -112,6 +121,10 @@ export default {
     value: {
       type: Object,
       default: () => {},
+    },
+    loading: {
+      type: Boolean,
+      default: false,
     },
   },
   data() {
@@ -188,7 +201,6 @@ export default {
     // Store accumulated changes
     editor.on('text-change', function (delta) {
       vm.change = vm.change.compose(delta);
-      // debouncedSave();
       vm.save();
       vm.$emit('input', vm.document);
     });
@@ -206,18 +218,6 @@ export default {
       const vm = this;
       if (vm.change && vm.change.length() > 0) {
         vm.document = JSON.parse(JSON.stringify(vm.editor.getContents()));
-
-        // vm.$fireStore
-        //   .collection('shoppinglists')
-        //   .doc(vm.$auth.user.uid)
-        //   .set(vm.document)
-        //   .then(function() {
-        //     const Delta = vm.editor.import('delta');
-        //     vm.change = new Delta();
-        //   })
-        //   .catch(function(error) {
-        //     console.error('Error writing document: ', error);
-        //   });
       }
       vm.$emit('submit');
     },
@@ -339,5 +339,9 @@ export default {
 
 span.ql-picker-label::before {
   color: #4dba87;
+}
+
+.ql-editor ul > li {
+  font-size: 24px;
 }
 </style>