+++ nuxt.config.js
@@ -43,7 +43,7 @@ const routes = (callback) => {
 
 const i18nSettings = {
   baseUrl: BASE_URL,
-  seo: true,
+  seo: false,
   defaultLocale: 'en',
   vueI18n: {
     fallbackLocale: 'en',
@@ -84,6 +84,11 @@ module.exports = {
     offlineAnalytics: true,
   },
 
+  meta: {
+    lang: undefined,
+    ogHost: 'https://pocketpasta.com',
+  },
+
   manifest: {
     name: pkg.name,
     short_name: 'pocketpasta',
@@ -101,7 +106,7 @@ module.exports = {
 
   router: {
     linkActiveClass: 'active',
-    middleware: ['theme'],
+    middleware: [],
   },
 
   env: {
@@ -135,23 +140,6 @@ module.exports = {
   },
 
   render: {
-    bundleRenderer: {
-      shouldPreload: (file, type) => {
-        // type is inferred based on the file extension.
-        // https://fetch.spec.whatwg.org/#concept-request-destination
-        if (type === 'script' || type === 'style') {
-          return true;
-        }
-        if (type === 'font') {
-          // only preload woff2 fonts
-          return /\.woff2$/.test(file);
-        }
-        // if (type === 'image') {
-        //   // only preload important images
-        //   return file === 'header-bg.jpg';
-        // }
-      },
-    },
     http2: {
       push: true,
       pushAssets: (req, res, publicPath, preloadFiles) =>
@@ -170,15 +158,8 @@ module.exports = {
       // If undefined or blank then we don't need the hyphen
       return titleChunk ? `${titleChunk} - PocketPasta` : 'PocketPasta';
     },
+    noscript: [{ innerHTML: 'This website requires JavaScript.', once: true }],
     meta: [
-      {
-        charset: 'utf-8',
-      },
-      {
-        name: 'viewport',
-        content:
-          'width=device-width, initial-scale=1, shrink-to-fit=no, minimal-ui',
-      },
       {
         property: 'og:title',
         template: (titleChunk) => {
@@ -188,38 +169,33 @@ module.exports = {
         vmid: 'og:title',
       },
       {
-        hid: 'description',
-        name: 'description',
-        content: pkg.description,
-      },
-      {
+        once: true,
         name: 'google-site-verification',
         content: 'LqVnUnYGR8NrvXrhnFgW5RjNJVChZp2j2OEP55xjE30',
       },
       {
+        once: true,
         name: 'version',
         content: pkg.version,
       },
       {
+        once: true,
         'http-equiv': 'Accept-CH',
         content: 'DPR, Viewport-Width, Width',
       },
     ],
     link: [
       {
+        once: true,
         rel: 'icon',
         type: 'image/x-icon',
         href: '/favicon.ico',
-      },
-      {
-        rel: 'preconnect',
-        href: 'https://res.cloudinary.com',
-        crossorigin: 'anonymous',
       },
       ...preconnectLinks.map((href) => ({
         rel: 'preconnect',
         href,
         crossorigin: 'anonymous',
+        once: true,
       })),
     ],
   },
@@ -244,16 +220,15 @@ module.exports = {
    */
   css: [
     '~/assets/scss/custom.scss',
-    '~/assets/css/font.css',
     '~/assets/css/custom.css',
-    'plyr/dist/plyr.css',
+    // 'plyr/dist/plyr.css',
   ],
 
   /*
    ** Plugins to load before mounting the App
    */
   plugins: [
-    { src: '~/plugins/vue-plyr' },
+    // { src: '~/plugins/vue-plyr' },
     // { src: '~/plugins/firebase.js', ssr: false },
     // { src: '~/plugins/firebase_auth.js', ssr: false },
     // { src: '~/plugins/quicklink', ssr: false },
@@ -264,6 +239,7 @@ module.exports = {
    ** Nuxt.js modules
    */
   modules: [
+    '@nuxt/content',
     '@nuxtjs/axios',
     '@nuxtjs/auth',
     '@nuxtjs/dotenv',
@@ -276,7 +252,6 @@ module.exports = {
     '@nuxtjs/sentry',
     'bootstrap-vue/nuxt',
     ['nuxt-i18n', i18nSettings],
-    'nuxt-webfontloader',
 
     // keep sitemap last
     '@nuxtjs/sitemap',
@@ -285,6 +260,7 @@ module.exports = {
   buildModules: [
     // Simple usage
     '@nuxtjs/vuetify',
+    '@nuxtjs/color-mode',
   ],
 
   dotenv: {
@@ -419,12 +395,6 @@ module.exports = {
       // priority: 0.5,
       lastmod: new Date(),
       // lastmodrealtime: true,
-    },
-  },
-
-  webfontloader: {
-    custom: {
-      families: ['Comic Neue'],
     },
   },